DOCS: fix 'object does not have a docstring'
