+++ sksurv/io/arffwrite.py
@@ -10,6 +10,10 @@
 #
 # You should have received a copy of the GNU General Public License
 # along with this program.  If not, see <http://www.gnu.org/licenses/>.
+"""
+Functions to write ARFF files.
+"""
+
 import os.path
 import re
