+++ sksurv/linear_model/coxnet.py
@@ -10,6 +10,10 @@
 #
 # You should have received a copy of the GNU General Public License
 # along with this program.  If not, see <http://www.gnu.org/licenses/>.
+"""
+Implementation of Cox's proportional hazards model with elastic-net regularization.
+"""
+
 import numbers
 import warnings
