+++ sksurv/svm/minlip.py
@@ -10,6 +10,10 @@
 #
 # You should have received a copy of the GNU General Public License
 # along with this program.  If not, see <http://www.gnu.org/licenses/>.
+"""
+Minlip variant implementations of the survival SVM.
+"""
+
 from abc import ABCMeta, abstractmethod
 import numbers
 import warnings
